backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-openai
openai
faiss-cpu
sentence-transformers
httpx[socks]
# httpcore[socks]
pypdf
cryptography>=3.1